Add profiler file & required changes
